+++ certbot/certbot/interfaces.py
@@ -254,6 +254,13 @@ class IConfig(zope.interface.Interface):
         "If updates provided by installer enhancements when Certbot is being run"
         " with \"renew\" verb should be disabled.")
 
+    preferred_chain = zope.interface.Attribute(
+        "If the CA offers multiple certificate chains for an issued certificate"
+        ", prefer the chain with an issuer matching this Subject Common Name or "
+        "Subject Key Identifier (hex-encoded). If no match, the default offered "
+        "chain will be used."
+    )
+
 class IInstaller(IPlugin):
     """Generic Certbot Installer Interface.